requirements.txt
PyYAML
requests
attrs